Reworked data.json loading and removed deprecated functions.
Hardcoding the data.json path in LocalDialogueAction was bad practice, instead I find this data in the configuration attribute that hovor loaded and add this into the config for the action. So no extra file read or path input is needed.

Since the old method of generating a response from required entities is worse than the new one, I removed it.
Since a lot of functions supported that and are also not longer needed, they were removed too.